Add NetcdfAdapterSSI
See LATIS-565.

This adapter uses a binary search on the domain variables to determine
what subset of the data to read from the NetCDF file rather than
reading all of it into memory. This should improve on speed and memory
usage compared to the normal NetcdfAdapter, but it currently assumes
the data being modeled is SSI-like.

There were a few hacks that were made to work around limitations due
to the adapter lifecycle. This implementation could use some cleaning
in general.