Fix MatchError for datasets without nested functions
See LISIRDIII-907.

We only need to update the length metadata when we're dealing with
nested functions. Here we deconstructed a Dataset to get to the nested
function, but this gave us a MatchError when there was no nested
function. If there is no nested function there is no length metadata
to update, so we can just return the Dataset as it was.

This still won't handle cases where we have a tuple and one or more of
the elements of that tuple are functions.