Add limit to NetcdfAdapter3
This commit adds a limit capability to `NetcdfAdapter3`.

This treats datasets as though they were uncurried, so the limit
represents the number of samples from the uncurried dataset. For
instance, a dataset modeled as `a -> b -> c` will be treated as `(a,
b) -> c` for the purposes of checking the limit.

Note that this limit applies to each granule individually.

See LATIS-669.